+++ packages/apps/loyalty-points/src/loyalty-points.ts
@@ -1,0 +1,12 @@
+import '@cloudcommerce/firebase/lib/init';
+import type { AppModuleBody } from '@cloudcommerce/types';
+import handleListPayments from './loyalty-list-payments';
+import handleCreateTransaction from './loyalty-create-transaction';
+
+export const listPayments = async (modBody: AppModuleBody) => {
+  return handleListPayments(modBody);
+};
+
+export const createTransaction = async (modBody: AppModuleBody) => {
+  return handleCreateTransaction(modBody);
+};