+++ packages/events/src/firebase.ts
@@ -15,4 +15,6 @@ export * from '@cloudcommerce/app-infinitepay/webhook';
 
 export * from '@cloudcommerce/app-galaxpay/events';
 
+export * from '@cloudcommerce/app-loyalty-points/events';
+
 export * from '@cloudcommerce/app-frenet/events';