+++ scripts/version-and-release.sh
@@ -4,4 +4,4 @@ pnpm build --force || exit 1
 pnpm test || exit 1
 pnpm run -r prerelease
 npx standard-version --commit-all
-npx zx scripts/release.mjs --publish --skip-stores
+npx zx scripts/release.mjs --publish